requirements.txt
GitPython==3.1.41
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
pathlib2==2.3.7; python_version < '3.4'
aiohttp==3.9.3
semgrep==1.63.0
ruff==0.3.2
chromadb==0.4.24
sentence-transformers==2.5.1
langgraph>=0.1.0
langchain-core>=0.2.0
litellm==1.31.2
pydantic==2.6.4
python-dotenv==1.0.1
numpy<2.0.0

# --- Phase Upgrade Dependencies (Master Plan) ---
# Task Orchestration (FastAPI + Celery + Redis)
fastapi>=0.111.0
uvicorn>=0.29.0
celery[redis]>=5.3.0
redis>=5.0.0

# ML Fusion Classifier (Phase 4.5)
scikit-learn>=1.4.0
shap>=0.44.0
joblib>=1.3.0
pandas>=2.2.0

# Ollama Health Check & HTTP Client
httpx>=0.27.0

# Dependency Scanning (Phase 9)
pip-audit>=2.7.0

# Testing
pytest>=8.0.0
